linux: Remove dirty-bitmap logic.
This is causing problems with 64bit PV guest.
Dirty-bitmap is intended to be used for backends that copy data between
the guest and displayed framebuffer, which is not what Surfman does by
default (drm-plugin will use m2p overrides in dom0 to let the GPU scan
the framebuffer of the guest after it has been mapped in dom0's address
space).

Remove as a work-around until a better solution is found.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1579